Make lookup and shutdown mutating
Motivation:
Feedback from API review

Modifications:
Make `lookup` and `shutdown` mutating, but run into problems with `PollingServiceDiscovery` and have to tweak it.

Without updating `PollingServiceDiscovery`:

```
cannot use mutating member on immutable value: 'self' is immutable
        self.lookup(service, deadline: nil) { result in
        ~~~~ ^
```

a. Add `mutating` to `PollingServiceDiscovery.subscribe` in the extension:

```
escaping closure captures mutating 'self' parameter
        self.lookup(service, deadline: nil) { result in
```

b. Make `PollingServiceDiscovery` class-only protocol:

```
cannot use mutating member on immutable value: 'self' is immutable
```

See:
- https://forums.swift.org/t/mutating-function-on-class-type-possible-bug/15066
- https://bugs.swift.org/browse/SR-142

In the end I make `PollingServiceDiscovery` class-only and redefine a non-mutating `lookup`. This is the most tolerable approach I can come up with.

Result:
`lookup` and `shutdown` are mutating.
