+++ Sources/ServiceDiscovery/ServiceDiscovery.swift
@@ -14,8 +14,6 @@
 
 import Dispatch
 
-// MARK: - Service discovery protocol
-
 /// Provides service instances lookup.
 public protocol ServiceDiscovery {
     /// Service identity type
@@ -35,7 +33,7 @@ public protocol ServiceDiscovery {
     /// Performs a lookup for the given service's instances. The result will be sent to `callback`.
     ///
     /// `defaultLookupTimeout` will be used to compute `deadline` in case one is not specified.
-    func lookup(_ service: Service, deadline: DispatchTime?, callback: @escaping (Result<[Instance], Error>) -> Void)
+    mutating func lookup(_ service: Service, deadline: DispatchTime?, callback: @escaping (Result<[Instance], Error>) -> Void)
 
     /// Subscribes to receive a service's instances whenever they change.
     ///
@@ -44,7 +42,7 @@ public protocol ServiceDiscovery {
     mutating func subscribe(to service: Service, handler: @escaping (Result<[Instance], Error>) -> Void)
 
     /// Performs clean up steps if any before shutting down.
-    func shutdown() throws
+    mutating func shutdown() throws
 }
 
 /// Errors that might occur during lookup.
@@ -78,46 +76,4 @@ public struct LookupError: Error, Equatable, CustomStringConvertible {
 
     /// Lookup has taken longer than allowed and thus has timed out.
     public static let timedOut = LookupError(type: .timedOut)
-}
-
-// MARK: - Polling service discovery protocol
-
-/// Polls for service instance updates at fixed interval.
-public protocol PollingServiceDiscovery: ServiceDiscovery {
-    /// The frequency at which `subscribe` will poll for updates.
-    var pollInterval: DispatchTimeInterval { get }
-}
-
-extension PollingServiceDiscovery {
-    public func subscribe(to service: Service, handler: @escaping (Result<[Instance], Error>) -> Void) {
-        self.lookup(service, deadline: nil) { result in
-            handler(result)
-
-            switch result {
-            case .success(let instances):
-                self._pollAndNotifyOnChange(service: service, previousInstances: instances, onChange: handler)
-            case .failure:
-                self._pollAndNotifyOnChange(service: service, previousInstances: nil, onChange: handler)
-            }
-        }
-    }
-
-    private func _pollAndNotifyOnChange(service: Service, previousInstances: [Instance]?, onChange: @escaping (Result<[Instance], Error>) -> Void) {
-        DispatchQueue.global().asyncAfter(deadline: DispatchTime.now() + self.pollInterval) {
-            guard !self.isShutdown else { return }
-
-            self.lookup(service, deadline: nil) { result in
-                // Subsequent lookups should only notify if instances have changed
-                switch result {
-                case .success(let instances):
-                    if previousInstances != instances {
-                        onChange(.success(instances))
-                    }
-                    self._pollAndNotifyOnChange(service: service, previousInstances: instances, onChange: onChange)
-                case .failure:
-                    self._pollAndNotifyOnChange(service: service, previousInstances: previousInstances, onChange: onChange)
-                }
-            }
-        }
-    }
 }
